django_project/requirements.txt
Django~=2.2.4
psycopg2-binary
djangorestframework
django-filter